requirements.txt
absl-py==2.1.0
accelerate==0.20.3
affine==2.4.0
aie==1.0
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
basemap==1.3.8
basemap-data==1.3.2
basemap-data-hires==1.3.2
bleach==6.0.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clip==1.0
cycler==0.11.0
datasets==2.13.2
decorator==5.1.1
dgl==1.1.3
diffusers==0.21.4
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.12.2
fiona==1.9.5
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gast==0.4.0
geopandas==0.10.2
geos==0.2.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.60.1
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
importlib-metadata==6.7.0
ipdb==0.13.13
ipykernel==5.5.5
ipython==7.34.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==5.3.4
jupyter_core==4.12.0
kaggle==1.5.16
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
loguru==0.7.2
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.4
multiprocess==0.70.14
networkx==2.6.3
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.535.133
oauthlib==3.2.2
open-clip-torch==2.23.0
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==22.3.1
prettytable==3.7.0
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.6
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pyecharts==2.0.4
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.2.1
pyshp==2.3.1
python-dateutil==2.8.2
python-slugify==8.0.1
python-version==0.0.2
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==19.0.2
rasterio==1.2.10
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rioxarray==0.9.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
segment-anything==1.0
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
setuptools==65.6.3
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.10
tokenizers==0.13.3
tomli==2.0.1
torch==1.11.0
torch-tb-profiler==0.4.3
torchaudio==0.11.0+cu113
torchtext==0.14.1
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
wandb==0.15.12
wcwidth==0.2.9
webencodings==0.5.1
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.16.0
xarray==0.20.2
xlrd==2.0.1
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.15.0
